% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmsy2catg.R
\name{bbmsy2catg}
\alias{bbmsy2catg}
\title{Convert B/BMSY to categorical status}
\usage{
bbmsy2catg(bbmsy, breaks = c(0.5, 1.5), catgs = c("over", "fully", "under"))
}
\arguments{
\item{bbmsy}{A vector, matrix, or dataframe of B/BMSY estimates.}

\item{breaks}{A vector of B/BMSY thresholds used to delineate status categories,
listed in order of ascending value (e.g., 0.2, 0.5, 1.5). Default is 0.5 and 1.5 where
B/BMSY < 0.5 = overexploited, 0.5 < B/BMSY < 1.5 = fully exploited, and B/BMSY > 1.5 = underexploited.}

\item{catgs}{A vector of names for the status categories, listed
in order of descending depletion (e.g, default is over, fully, under).}
}
\value{
A vector, matrix, or dataframe of status categories.
}
\description{
Converts B/BMSY estimates to categorical statuses (e.g., underexploited,
fully exploited, or overexploited).
}
\examples{
bbmsy <- select(preds, bbmsy,  mprm, comsir, sscom, cmsy13, cmsy17, zbrt, ocom, super1)
status <- bbmsy2catg(bbmsy)
}
