% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bbmsy.R
\name{bbmsy2s}
\alias{bbmsy2s}
\title{Convert saturation to B/BMSY}
\usage{
bbmsy2s(bbmsy)
}
\arguments{
\item{bbmsy}{B/BMSY values(s)}
}
\value{
Saturation (B/K) value(s)
}
\description{
Converts saturation (S = B/K) to B/BMSY
}
\details{
S = 1 - depletion = B / K = B/BMSY / 2
B/BMSY = S * 2
}
