% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbrt.R
\name{catchParam}
\alias{catchParam}
\title{Compute predictors for the Zhou-BRT catch-only stock assessment model}
\usage{
catchParam(catchData)
}
\arguments{
\item{catchData}{A dataframe with columns "stock", "yr", and "catch"}
}
\description{
This function calculates the catch statistics used as predictors in the Zhou-BRT
catch-only stock assessment model.
}
