% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsy2.R
\name{cmsy2}
\alias{cmsy2}
\title{cMSY catch-only stock assessment model}
\usage{
cmsy2(year, catch, resilience, r.low = NA, r.hi = NA, stb.low = NA,
  stb.hi = NA, int.yr = NA, intb.low = NA, intb.hi = NA,
  endb.low = NA, endb.hi = NA, q.start = NA, q.end = NA, verbose = T)
}
\arguments{
\item{year}{A time series of years}

\item{catch}{A time series of catch}

\item{resilience}{Resilience of the stock: "high", "medium", "low", "very low", "unknown"}

\item{int.yr}{A user-specified year of intermediate biomass (optional)}

\item{verbose}{Set to FALSE to suppress printed updates on CMSY/BSM progress (default=TRUE)}

\item{r.low/r.hi}{A user-specified prior on the species intrinsic growth rate, r (optional)}

\item{stb.low/stb.hi}{A user-specified prior on biomass relative to unfished biomass at the beginning of the catch time series (optional)}

\item{intb.low/intb.hi}{A user-specified prior on biomass relative to unfished biomass in the intermediate year (optional)}

\item{endb.low/endb.hi}{A user-specified prior on biomass relative to unfished biomass at the end of the catch time series (optional)}

\item{q.start/q.end}{A user-specified start and end year for estimating the catchability coefficient (optional; default is last 5 years)}
}
\value{
A time series of B/BSMY estimates
}
\description{
Estimates B/BMSY and biological quantities and reference points using only
a time series of catch and estimate of resilience using cMSY from Froese et al. (2017).
}
\examples{
output <- cmsy2(year=SOLIRIS$yr, catch=SOLIRIS$ct, r.low=0.18, r.hi=1.02)
plot_cmsy2(output)
plot_cmsy2_mgmt(output)
}
\references{
Froese R, Demirel N, Coro G, Kleisner KM, Winker H (2017)
Estimating fisheries reference points from catch and resilience. Fish and Fisheries 18(3): 506-526.
\url{http://onlinelibrary.wiley.com/doi/10.1111/faf.12190/abstract}
}
