% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbrt.R
\name{fit_zbrt}
\alias{fit_zbrt}
\title{Fit the Zhou-BRT catch-only stock assessment model}
\usage{
fit_zbrt(yr, catch)
}
\arguments{
\item{yr}{A time series of years}

\item{catch}{A time series of catch}
}
\value{
A time series of B/BMSY estimates
}
\description{
This function fits the Zhou-BRT catch-only stock assessment model. It
only requires a time series of catch.
}
