% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_cmsy.R
\name{ms_cmsy}
\alias{ms_cmsy}
\title{Multispecies cMSY (beta version)}
\usage{
ms_cmsy(data, key, npairs = 5000)
}
\arguments{
\item{data}{A dataframe with the following columns: stock, year, catch}

\item{key}{A dataframe with the following columns:
\describe{
\item{stock}{Name of the stock}
\item{family}{Taxanomic family of each stock}
\item{resilience}{Resilience of each stock (i.e., very low, low, medium, high)}
\item{id_fixed}{A boolean to indicate whether initial depletion should be estimated (FALSE) or fixed at zero (TRUE).}
}}

\item{npairs}{Number of r-K pairs to evaluate (default=5,000)}
}
\value{
A list with the following elements:
\item{stocks}{Names of the analyzed stocks/species}
\item{yrs}{Years represented in the time series}
\item{r_priors}{R priors for each species}
\item{k_priors}{K priors for each species}
\item{s1_priors}{Initial saturation priors for each species}
\item{s2_priors}{Final saturation priors for each species}
\item{id_try}{Initial saturation values evaluated}
\item{r_try}{R values evaluated}
\item{k_try}{K values evaluated}
\item{id_rk_v}{Viable r-K-initial saturation combos}
\item{b_v}{Viable biomass trajectories}
\item{bbmsy_v}{Viable B/BMSY trajectories}
\item{bbmsy_v_median}{Median of viable B/BMSY trajectories (cMSY estimate of status)}
\item{er_v}{Viable exploitation rate trajectories}
\item{er_vv}{Most probable exploitation rate trajectories based on effort constraint}
\item{top_corr}{Most highly correlated r-k-initial saturation combos from each row of correlation matrix}
\item{bbmsy_vv}{Most probable B/BMSY trajectories based on effort constraint}
\item{bbmsy_vv_median}{Median of most probable B/BMSY trajectories (MS-cMSY estimate of status)}
\item{method}{Name of the method}
}
\description{
Estimates stock status (B/BMSY) time series for multispecies fisheries from
time series of catch uing the multispecies cMSY (MS-cMSY) method of
Free et al. (in prep). Note: this model is still under development.
}
\details{
Catch-only single-species stock assessment methods are generally poor predictors of stock status because changes in catch can reflect both changes in population size and fishing effort. However, when multiple species are harvested by a single fleet, each species may share the same underlying effort dynamics, which places an informative constraint on the interpretation of the catch time series. We developed a multispecies catch-only stock assessment model, MS-cMSY, that leverages these principles to estimate stock status. MS-cMSY is a multispecies extension of cMSY that identifies viable population trajectories as those that maximize correlation in implied fishing effort trends.
}
\examples{
# Example coming soon
a <- c(1,2)
}
\references{
Free CM, Rudd MB, Kleisner KM, Thorson JT, Longo C, Minto C,
Jensen OP (in prep) Multispecies catch-only models for assessing data-limited fisheries.
}
