% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zbrt.R
\name{plot_zbrt}
\alias{plot_zbrt}
\title{Plot zBRT results}
\usage{
plot_zbrt(output)
}
\arguments{
\item{output}{Output from the zBRT model (see ?zbrt)}
}
\value{
Plots
}
\description{
Plot zBRT results
}
\examples{
output <- zbrt(year=YELLSNEMATL$year, catch=YELLSNEMATL$tc)
plot_zbrt(output)
}
