% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preds}
\alias{preds}
\title{RAM Legacy Database catch-only status predictions}
\format{A data frame with 161 rows (stocks) and 57 variables including B/BMSY estimates
and status estimates.}
\usage{
preds
}
\description{
A dataset containing the catch and biomass time series for USA Southern
New England/Mid-Atlantic (SNE/MA) Yellowtail flounder (\emph{Pleuronectes ferruginea}) from 1973-2014.
This dataset is included for additional testing of the \pkg{datalimited2} package.
}
\keyword{datasets}
