% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience.R
\name{resilience}
\alias{resilience}
\title{Retrieve resilience estimates from FishBase}
\usage{
resilience(species)
}
\arguments{
\item{species}{A character vector of species}
}
\value{
A dataframe containing species and their resilience values
}
\description{
Retrieves estimates of resilience for fish and invertebrate species from FishBase
and SeaLifeBase. Resilience describes the ability of a species population to recover
after a perturbance and is reported as: very low, low, medium, or high.
}
\examples{
# Retrieve resilience for three species
resilience(species=c("Gadus morhua", "Mola mola", "Centropristis striata"))
}
