% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TIGERFLAT}
\alias{TIGERFLAT}
\title{SE Australia Tiger flathead time series}
\format{A data frame with 98 rows (years) and 3 variables:
\describe{
  \item{Stock}{stock id}
  \item{yr}{year}
  \item{ct}{catch, in metric tons}
}}
\source{
Zhou S, Punt AE, Smith ADM, Ye Y, Haddon M, Dichmont CM, Smith DC
(2017) An optimised catch-only assessment method for data poor fisheries.
ICES Journal of Marine Science: doi:10.1093/icesjms/fsx226.
\url{https://doi.org/10.1093/icesjms/fsx226}
}
\usage{
TIGERFLAT
}
\description{
A dataset containing the catch time series for SE Australia Tiger flathead
(Neoplatycephalus richardsoni) from 1915-2012. This stock was used as an example in the OCOM
paper and is used to validate this package's implementation of zBRT and OCOM.
}
\keyword{datasets}
