% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YELLSNEMATL}
\alias{YELLSNEMATL}
\title{USA SNE/MA Yellowtail flounder time series}
\format{A data frame with 42 rows (years) and XX variables:
\describe{
  \item{Stock}{stock id}
  \item{yr}{year}
  \item{ct}{catch, in metric tons}
  ...
}}
\usage{
YELLSNEMATL
}
\description{
A dataset containing the catch time series for USA SNE/MA Yellowtail flounder
(Pleuronectes ferruginea) from 1973-2014.
}
\keyword{datasets}
