% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbrt.R
\name{zbrt}
\alias{zbrt}
\title{Zhou-BRT catch-only stock assessment model}
\usage{
zbrt(year, catch)
}
\arguments{
\item{year}{A time series of years}

\item{catch}{A time series of catch}
}
\value{
A dataframe with a time series of saturation and B/BMSY estimates. S8 and S38
correspond to the saturation estimates from the 8- and 38-predictor models, respectively.
S, the best estimate of saturation, is the mean of these two predictions.
B/BMSY is this estimate doubled (B/BMSY = S * 2). High and low values correspond to the
upper and lower 95% confidence intervals, respectively.
}
\description{
Estimates saturation (B/k) and stock status (B/BMSY) time series from a
time series of catch using the boosted regression tree (BRT) model from Zhou et al. (2017).
}
\examples{
output <- zbrt(year=TIGERFLAT$yr, catch=TIGERFLAT$catch)
plot_dlm(output)
}
\references{
Zhou S, Punt AE, Yimin Y, Ellis N, Dichmont CM, Haddon M, Smith DC, Smith ADM
(2017) Estimating stock depletion level from patterns of catch history. \emph{Fish and Fisheries}.
\url{http://onlinelibrary.wiley.com/doi/10.1111/faf.12201/abstract}
}
