% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbrt.R
\name{catchParam}
\alias{catchParam}
\title{Compute predictors for the Zhou-BRT catch-only stock assessment model}
\usage{
catchParam(catchData)
}
\description{
Compute predictors for the Zhou-BRT catch-only stock assessment model
}
