% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsy2.R
\name{cmsy2}
\alias{cmsy2}
\title{CMSY catch-only stock assessment model}
\usage{
cmsy2(year, catch, resilience = NA, r.low = NA, r.hi = NA, stb.low = NA,
  stb.hi = NA, int.yr = NA, intb.low = NA, intb.hi = NA,
  endb.low = NA, endb.hi = NA, verbose = T)
}
\arguments{
\item{year}{A time series of years}

\item{catch}{A time series of catch}

\item{resilience}{Resilience of the stock: "High", "Medium", "Low", or "Very low" (optional if \code{r.low} and \code{r.hi} are specified)}

\item{r.low, r.hi}{A user-specified prior on the species intrinsic growth rate, r (optional if \code{resilience} is specified)}

\item{stb.low, stb.hi}{A user-specified prior on biomass relative to unfished biomass at the beginning of the catch time series (optional)}

\item{int.yr}{A user-specified year of intermediate biomass (optional)}

\item{intb.low, intb.hi}{A user-specified prior on biomass relative to unfished biomass in the year of intermediate biomass (optional)}

\item{endb.low, endb.hi}{A user-specified prior on biomass relative to unfished biomass at the end of the catch time series (optional)}

\item{verbose}{Set to FALSE to suppress printed updates on model progress (default=TRUE)}
}
\value{
A list with the following elements:
\item{ref_pts}{A data frame with biological quantity and reference point estimates with 95\% confidence intervals}
\item{ref_ts}{A data frame with B/BMSY, F/FMSY, biomass, and fishing mortality time series with 95\% confidence intervals}
\item{priors}{A data frame with the priors used in the analysis}
\item{r_viable}{A vector with the viable r values}
\item{k_viable}{A vector with the viable K values}
\item{bt_viable}{A data frame with the biomass trajectories produced by the viable r/K pairs}
\item{method}{Name of the method}
}
\description{
Estimates biomass, fishing mortality, and stock status (i.e., B/BMSY, F/FMSY)
time series and biological/management quantities (i.e., r, K, MSY, BMSY, FMSY)
from a time series of catch and a resilience estimate using CMSY from Froese et al. 2017.
}
\details{
The CMSY model developed by Froese et al. 2017 employs a stock
reduction analysis using priors for r based on resilience, K based on
maximum catch and the r priors, and start, intermediate, and final year
saturation based on a set of simple rules. It also allows
users to revise the default priors based on expert knowledge. The SRA employs
a Schaefer biomass dynamics model and an algorithm for identifying feasible
parameter combinations to estimate biomass, fishing mortality, and stock status
(i.e., B/BMSY, F/FMSY) time series and biological/management quantities
(i.e., r, K, MSY, BMSY, FMSY).

This implementation of CMSY is based on "version q". CMSY is constantly being improved
so please consult the original authors about where to find the most up-to-date method.
}
\examples{
# Fit cMSY to catch time series and plot output
output <- cmsy2(year=SOLIRIS$yr, catch=SOLIRIS$ct, r.low=0.18, r.hi=1.02)
plot_dlm(output)

# Extract reference points and time series from output
ref_pts <- output[["ref_pts"]]
ref_ts <- output[["ref_ts"]]
}
\references{
Froese R, Demirel N, Coro G, Kleisner KM, Winker H (2017)
Estimating fisheries reference points from catch and resilience. \emph{Fish and Fisheries} 18(3): 506-526.
\url{http://onlinelibrary.wiley.com/doi/10.1111/faf.12190/abstract}
}
