% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsy2.R
\name{cmsy2}
\alias{cmsy2}
\title{cMSY catch-only stock assessment model}
\usage{
cmsy2(year, catch, resilience = NA, r.low = NA, r.hi = NA, stb.low = NA,
  stb.hi = NA, int.yr = NA, intb.low = NA, intb.hi = NA,
  endb.low = NA, endb.hi = NA, q.start = NA, q.end = NA, verbose = T)
}
\arguments{
\item{year}{A time series of years}

\item{catch}{A time series of catch}

\item{resilience}{Resilience of the stock: "High", "Medium", "Low", "Very low"}

\item{r.low, r.hi}{A user-specified prior on the species intrinsic growth rate, r (optional)}

\item{stb.low, stb.hi}{A user-specified prior on biomass relative to unfished biomass at the beginning of the catch time series (optional)}

\item{int.yr}{A user-specified year of intermediate biomass (optional)}

\item{intb.low, intb.hi}{A user-specified prior on biomass relative to unfished biomass in the intermediate year (optional)}

\item{endb.low, endb.hi}{A user-specified prior on biomass relative to unfished biomass at the end of the catch time series (optional)}

\item{q.start, q.end}{A user-specified start and end year for estimating the catchability coefficient (optional; default is last 5 years)}

\item{verbose}{Set to FALSE to suppress printed updates on CMSY/BSM progress (default=TRUE)}
}
\value{
A list containing the following elements:
(1) ref_pts - A dataframe with biological quantity / reference point estimates with 95% confidence intervals;
(2) ref_ts - A dataframe with B/BMSY and reference point time series with 95% confidence intervals;
(3) priors - A dataframe with the priors used in the cMSY analysis;
(4) rv.all - A vector with the viable r values;
(5) kv.all - A vector with the viable k values;
(6) btv.all - A dataframe with the biomass trajectories produced by the viable r/k pairs.
}
\description{
Estimates B/BMSY time series and other biological quantities (e.g., r, k, MSY) from
a time series of catch and a resilience estimate using cMSY from Froese et al. (2017).
}
\examples{
output <- cmsy2(year=SOLIRIS$yr, catch=SOLIRIS$ct, r.low=0.18, r.hi=1.02)
plot_dlm(output)
}
\references{
Froese R, Demirel N, Coro G, Kleisner KM, Winker H (2017)
Estimating fisheries reference points from catch and resilience. Fish and Fisheries 18(3): 506-526.
\url{http://onlinelibrary.wiley.com/doi/10.1111/faf.12190/abstract}
}
