% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocom.R
\name{ocom}
\alias{ocom}
\title{Optimized catch-only model}
\usage{
ocom(year, catch, m)
}
\arguments{
\item{year}{A time series of years}

\item{catch}{A time series of catch}

\item{m}{Natural mortality (1/yr)}
}
\value{
A list with the following elements:
\item{ref_pts}{A data frame with estimates of r, k, MSY, BMSY, FMSY and final year saturation and confidence intervals}
\item{ref_ts}{A data frame with time series of biomass, saturation, fishing mortality, B/BMSY, and F/FMSY estimates and 95\% confidence intervals}
\item{b_ts}{A data frame with time series of biomass estimates and expanded confidence intervals}
\item{s_ts}{A data frame with time series of saturation estimates and expanded confidence intervals}
\item{f_ts}{A data frame with time series of fishing mortality (F) estimates and expanded confidence intervals}
\item{bbmsy_ts}{A data frame with time series of B/BMSY estimates and expanded confidence intervals}
\item{ffmsy_ts}{A data frame with time series of F/FMSY estimates and expanded confidence intervals}
\item{b_trajs}{A data frame with 1000 randomly selected biomass trajectories}
\item{s_trajs}{A data frame with 1000 corresponding saturation trajectories}
\item{f_trajs}{A data frame with 1000 corresponding fishing mortality (F) trajectories}
\item{bbmsy_trajs}{A data frame with 1000 corresponding B/BMSY trajectories}
\item{ffmsy_trajs}{A data frame with 1000 corresponding F/FMSY trajectories}
\item{krms_draws}{A data frame with 10,000 draws underpinning the estimakes of r, k, MSY, and saturation}
\item{method}{Name of the method}
}
\description{
Estimates biomass, fishing mortality, and stock status (B/BMSY, F/FMSY) time series and
biological/management quantities (i.e., r, k, MSY, BMSY, FMSY) from a time
series of catch and a natural mortality (M) estimate using the optimized
catch-only model (OCOM) from Zhou et al. 2017.
}
\details{
The "optimized catch-only model" (OCOM) developed by Zhou et al. 2017
employs a stock reduction analysis (SRA) using priors for r and stock depletion
derived from natural mortality and saturation estimated using the Zhou-BRT method, respectively.
The SRA employs a Schaefer biomass dynamics model and an algorithm for identifying
feasible parameter combinations to estimate biomass, fishing mortality, and stock status (B/BMSY, F/FMSY) time series and
biological/management quantities (i.e., r, K, MSY, BMSY, FMSY).
}
\examples{
# Fit OCOM to catch time series and plot output
set.seed(1) # stochastic fitting
output <- ocom(year=TIGERFLAT$yr, catch=TIGERFLAT$catch, m=0.27)
plot_dlm(output)

# Extract reference points and time series from output
ref_pts <- output[["ref_pts"]]
ref_ts <- output[["ref_ts"]]
}
\references{
Zhou S, Punt AE, Smith ADM, Ye Y, Haddon M, Dichmont CM, Smith DC
(2017) An optimised catch-only assessment method for data poor fisheries.
\emph{ICES Journal of Marine Science}: doi:10.1093/icesjms/fsx226.
\url{https://doi.org/10.1093/icesjms/fsx226}
}
