% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocom.R
\name{ocom}
\alias{ocom}
\title{Optimized catch-only model}
\usage{
ocom(year, catch, m)
}
\arguments{
\item{year}{A time series of years}

\item{catch}{A time series of catch}

\item{m}{Natural mortality (1/yr)}
}
\value{
A list of length nine containing the following elements:
(1) a time series of biomass estimates;
(2) a time series of saturation estimates;
(3) a time series of B/BMSY estimates;
(4) 1000 randomly selected biomass trajectories;
(5) 1000 corresponding saturation trajectories;
(6) 1000 corresponding B/BMSY trajectories;
(7) the estimates of biological quanties r, k, MSY, and S;
(8) 10,000 draws underpinning these values; and
(9) the name of the method.
}
\description{
Estimates saturation (B/K) and stock status (B/BMSY) time series and
other biological quantities (i.e., r, k, MSY, final year saturation) from a time
series of catch and natural mortality (M) estimate using the optimized
catch-only model (OCOM) from Zhou et al. 2017.
}
\details{
The "optimized catch-only model" (OCOM) developed by Zhou et al. 2017
employs a stock reduction analysis (SRA) using priors for r and stock depletion
derived from natural mortality and saturation estimated by the Zhou-BRT method, respectively.
The SRA uses a Schaefer biomass dynamics model and an algorithm for identifying
feasible parameter combinations to estimate biological quantities such as B0, r,
annual biomass, and depletion as well as management quantities such as MSY, BMSY, and FMSY.
}
\examples{
# Fit OCOM to catch time series and plot output
set.seed(1) # stochastic fitting
output <- ocom(year=TIGERFLAT$yr, catch=TIGERFLAT$catch, m=0.27)
plot_dlm(output)

# Extract reference points and time series from output
b_ts <- output[["b_ts"]]
bbmsy_ts <- output[["bbmsy_ts"]]
krms <- output[["krms"]]
}
\references{
Zhou S, Punt AE, Smith ADM, Ye Y, Haddon M, Dichmont CM, Smith DC
(2017) An optimised catch-only assessment method for data poor fisheries.
ICES Journal of Marine Science: doi:10.1093/icesjms/fsx226.
\url{https://doi.org/10.1093/icesjms/fsx226}
}
