% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocom.R
\name{ocom}
\alias{ocom}
\title{Optimized catch-only model}
\usage{
ocom(year, catch, m)
}
\arguments{
\item{year}{A time series of years}

\item{catch}{A time series of catch}

\item{m}{Natural mortality (1/yr)}
}
\value{
A list with the following elements: (1) time series of B/BMSY estimates;
(2) 1000 randomly selected biomass trajectories; (3) 1000 corresponding B/BMSY
trajectories; (4) estimates of biological quanties r, k, MSY, S; and (5) the
10,000 draws underpinning these values.
}
\description{
Estimates biological quantities (r, k, MSY, final year saturation) and
stock status (B/BMSY) from a time series of catch and natural mortality (M)
estimate using the optimized catch-only model from Zhou et al. (2017).
}
\examples{
ocom(year=YELLSNEMATL$year, catch=YELLSNEMATL$tc, m=0.2)
}
\references{
Zhou S, Punt AE, Smith ADM, Ye Y, Haddon M, Dichmont CM, Smith DC
(2017) An optimised catch-only assessment method for data poor fisheries.
ICES Journal of Marine Science: doi:10.1093/icesjms/fsx226.
\url{https://doi.org/10.1093/icesjms/fsx226}
}
