% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmsy2.R
\name{plot_cmsy2}
\alias{plot_cmsy2}
\title{Plot cMSY model results}
\usage{
plot_cmsy2(output)
}
\arguments{
\item{output}{Output from the cMSY stock assessment model (see ?cmsy2)}
}
\value{
Six plots: (1) catch time series; (2) r-k pair search; (3) viable r-k pairs;
(4) saturation time series; (5) exploitation rate time series; (6) surplus production curve
}
\description{
Plots cMSY model results for following the example of Froese et al. (2016).
The following six plots are produced:
\itemize{
  \item{A - Catch time series}
  \item{B - Finding r-k pairs}
  \item{C - Viable r-k pairs}
  \item{D - Saturation (B/k) time series}
  \item{E - Exploitation rate (F / (r/2)) time series}
  \item{F - Surplus production curve}
}
}
\examples{
output <- cmsy2(year=SOLIRIS$yr, catch=SOLIRIS$ct, r.low=0.18, r.hi=1.02)
plot_cmsy2(output)
plot_cmsy2_mgmt(output)
}
\references{
Froese R, Demirel N, Coro G, Kleisner KM, Winker H (2016)
A Simple User Guide for CMSY and BSM (version “q”). 27 October 2016.
\url{http://oceanrep.geomar.de/33076/}
}
