% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmsy2_mgmt.R
\name{plot_cmsy2_mgmt}
\alias{plot_cmsy2_mgmt}
\title{Plot cMSY and BSM model results for management}
\usage{
plot_cmsy2_mgmt(output)
}
\arguments{
\item{output}{Output from the cMSY or BSM stock assessment model (see ?cmsy2 or ?bsm)}
}
\value{
Four plots: (1) catch time series; (2) B/BMSY time series; (3) F/FMSY time serie; and (4) Kobe plot
}
\description{
Plots cMSY and BSM model results for management following the example of Froese et al. (2016).
Produces the following four plots:
\itemize{
  \item{A - Catch time series}
  \item{B - B/BMSY time series}
  \item{C - F/FMSY time series}
  \item{D - Kobe plot}
}
}
\examples{
# Fit cMSY and plot results
output <- cmsy2(year=SOLIRIS$yr, catch=SOLIRIS$ct, r.low=0.18, r.hi=1.02)
plot_cmsy2(output)
plot_cmsy2_mgmt(output)

# Fit BSM and plot results
output <- bsm(year=SOLIRIS$yr, catch=SOLIRIS$ct, biomass=SOLIRIS$bt, btype="CPUE", r.low=0.18, r.hi=1.02)
plot_cmsy2(output)
plot_cmsy2_mgmt(output)
}
\references{
Froese R, Demirel N, Coro G, Kleisner KM, Winker H (2016)
A Simple User Guide for CMSY and BSM (version “q”). 27 October 2016.
\url{http://oceanrep.geomar.de/33076/}
}
