% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorcs.R
\name{rorcs}
\alias{rorcs}
\title{Predict stock status using the refined ORCS approach}
\usage{
rorcs(scores)
}
\arguments{
\item{scores}{A numeric vector of length twelve containing scores for the
following "Table of Attributes" questions:
\itemize{
  \item{TOA 1 - Status of assessed stocks in fishery}
  \item{TOA 3 - Behavior affecting capture (2 or 3 only)}
  \item{TOA 5 - Discard rate}
  \item{TOA 6 - Targeting intensity}
  \item{TOA 7 - M compared to dominant species}
  \item{TOA 8 - Occurence in catch}
  \item{TOA 9 - Value (US$/lb) - continuous value}
  \item{TOA 10 - Recent trends in catch}
  \item{TOA 11 - Habitat loss}
  \item{TOA 12 - Recent trend in effort}
  \item{TOA 13 - Recent trend in abundance index}
  \item{TOA 14 - Proportion of population protected}
}}
}
\value{
Stock status (i.e., under, fully, or overexploited)
}
\description{
This function predicts stock status (i.e., under, fully, or overexploited) using
the refined ORCS approach (Free et al. 2017).
}
\references{
Free CM, Jensen OP, Wiedenmann J, Deroba JJ (2017) The
refined ORCS approach: a catch-based method for estimating stock status
and catch limits for data-poor fish stocks. Fisheries Research 193: 60-70.
\url{https://doi.org/10.1016/j.fishres.2017.03.017}
}
