% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bbmsy.R
\name{s2bbmsy}
\alias{s2bbmsy}
\title{Convert B/BMSY to saturation}
\usage{
s2bbmsy(s)
}
\arguments{
\item{s}{Saturation (B/K) value(s)}
}
\value{
B/BMSY values(s)
}
\description{
Converts B/BMSY to saturation (S = B/K).
}
\details{
S = 1 - depletion = B / K = B/BMSY / 2
B/BMSY = S * 2
}
