% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbrt.R
\name{zbrt}
\alias{zbrt}
\title{Fit the Zhou-BRT catch-only stock assessment model}
\usage{
zbrt(yr, catch)
}
\arguments{
\item{yr}{A time series of years}

\item{catch}{A time series of catch}
}
\value{
B/BMSY estimate for the terminal year of the time series
}
\description{
This function fits the Zhou-BRT catch-only stock assessment model. It
only requires a time series of catch.
}
\references{
Zhou S, Punt AE, Yimin Y, Ellis N, Dichmont CM, Haddon M, Smith DC, Smith ADM
(2017) Estimating stock depletion level from patterns of catch history. Fish and Fisheries.
\url{http://onlinelibrary.wiley.com/doi/10.1111/faf.12201/abstract}
}
