% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbrt.R
\name{zbrt}
\alias{zbrt}
\title{Zhou-BRT catch-only stock assessment model}
\usage{
zbrt(year, catch)
}
\arguments{
\item{year}{A time series of years}

\item{catch}{A time series of catch}
}
\value{
A list with the following elements:
\item{ts}{A data frame with a time series of saturation and B/BMSY estimates.
S8 and S38 correspond to the saturation estimates from the 8- and 38-predictor models, respectively.
S, the best estimate of saturation, is the mean of these two predictions.
B/BMSY is this estimate doubled (B/BMSY = S * 2). High and low values correspond to the
upper and lower 95\% confidence intervals, respectively.}
\item{method}{Name of the method}
}
\description{
Estimates saturation (B/K) and stock status (B/BMSY) time series from a
time series of catch using the boosted regression tree (BRT) model from Zhou et al. 2017.
Note: B/BMSY is equal to saturation times two.
}
\details{
Zhou et al. 2017 use boosted regression tree models (Zhou-BRT)
trained on the RAM Legacy Database to estimate saturation (i.e., 1 - depletion = 0.5*B/BMSY)
from 56 catch history statistics, the most important of which are linear
regression coefficients for the whole catch time series, the subseries before
and after the maximum catch, and in recent years. Ultimately, saturation is
estimated as the average of the saturation values predicted by two reduced
and bias-corrected BRT models (8 and 38 predictors each). B/BMSY is estimated as saturation doubled.
}
\examples{
# Fit zBRT model to catch time series and plot output
output <- zbrt(year=TIGERFLAT$yr, catch=TIGERFLAT$catch)
plot_dlm(output)

# Extract time series from output
ts <- output[["ts"]]
}
\references{
Zhou S, Punt AE, Yimin Y, Ellis N, Dichmont CM, Haddon M, Smith DC, Smith ADM
(2017) Estimating stock depletion level from patterns of catch history. \emph{Fish and Fisheries}.
\url{http://onlinelibrary.wiley.com/doi/10.1111/faf.12201/abstract}
}
